# api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AudiModal.ai API",
    "description": "Comprehensive AI-powered document processing and analytics platform with ML/AI classification, anomaly detection, unified sync framework, and enterprise integrations.",
    "version": "1.0.0",
    "contact": {
      "name": "AudiModal.ai",
      "url": "https://audimodal.ai",
      "email": "support@audimodal.ai"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "https://api.audimodal.ai/v1",
      "description": "Production server"
    },
    {
      "url": "https://staging-api.audimodal.ai/v1",
      "description": "Staging server"
    },
    {
      "url": "http://localhost:8080/v1",
      "description": "Development server"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    },
    {
      "ApiKeyAuth": []
    }
  ],
  "paths": {
...
    "/anomalies": {
      "get": {
        "tags": ["Anomaly Detection"],
        "summary": "List anomalies",
        "description": "Retrieve detected anomalies with filtering options",
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "description": "Filter by anomaly type",
            "schema": {
              "type": "string",
              "enum": ["statistical", "content", "behavioral", "security"]
            }
          },
          {
            "name": "severity",
            "in": "query",
            "description": "Filter by severity level",
            "schema": {
              "type": "string",
              "enum": ["low", "medium", "high", "critical"]
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Filter by anomaly status",
            "schema": {
              "type": "string",
              "enum": ["new", "investigating", "resolved", "false_positive"]
            }
          },
          {
            "name": "start_time",
            "in": "query",
            "description": "Filter anomalies after this time",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "end_time",
            "in": "query",
            "description": "Filter anomalies before this time",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Anomalies retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "anomalies": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Anomaly"
                      }
                    },
                    "total": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/anomalies/{anomalyId}": {
      "get": {
        "tags": ["Anomaly Detection"],
        "summary": "Get anomaly",
        "description": "Retrieve details of a specific anomaly",
        "parameters": [
          {
            "name": "anomalyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Anomaly details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Anomaly"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["Anomaly Detection"],
        "summary": "Update anomaly status",
        "description": "Update the status or add notes to an anomaly",
        "parameters": [
          {
            "name": "anomalyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": ["new", "investigating", "resolved", "false_positive"]
                  },
                  "notes": {
                    "type": "string"
                  },
                  "assigned_to": {
                    "type": "string",
                    "format": "uuid"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Anomaly updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Anomaly"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    },
    "schemas": {
...
      "Anomaly": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "tenant_id": {
            "type": "string",
            "format": "uuid"
          },
          "type": {
            "type": "string",
            "enum": ["statistical", "content", "behavioral", "security"]
          },
          "severity": {
            "type": "string",
            "enum": ["low", "medium", "high", "critical"]
          },
          "status": {
            "type": "string",
            "enum": ["new", "investigating", "resolved", "false_positive"]
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "confidence": {
            "type": "number",
            "format": "float"
          },
          "score": {
            "type": "number",
            "format": "float"
          },
          "data": {
            "type": "object",
            "properties": {
              "entity_id": {
                "type": "string"
              },
              "entity_type": {
                "type": "string"
              },
              "metrics": {
                "type": "object"
              },
              "context": {
                "type": "object"
              }
            }
          },
          "threshold": {
            "type": "object",
            "properties": {
              "expected_value": {
                "type": "number",
                "format": "float"
              },
              "threshold_value": {
                "type": "number",
                "format": "float"
              },
              "deviation": {
                "type": "number",
                "format": "float"
              }
            }
          },
          "detected_at": {
            "type": "string",
            "format": "date-time"
          },
          "resolved_at": {
            "type": "string",
            "format": "date-time"
          },
          "assigned_to": {
            "type": "string",
            "format": "uuid"
          },
          "notes": {
            "type": "string"
          },
          "related_entities": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "entity_id": {
                  "type": "string"
                },
                "entity_type": {
                  "type": "string"
                },
                "relationship": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
...
    }
  },
...
}